Final quat fix (for now)
The optimal method is to store rotations as angles within the transform (which we are already doing). This means we can use euler to make transformations without having to convert a quat to euler angles (which causes issues). However, bullet only stores quats, which means when an entity has a RBC, their angles have to be set by a quat. This means that the euler angles appear sporadic, especially when dealing with them in scripts or the properties panel. This isn't something that can easily be fixed (if at all), so instead I added an ApplyRotation function that will apply a delta rotation to the current rotation regardless of the current orientation of the object. This becomes necessary when trying to modify the rotations of physics objects as the Y rotation is locked between -90 and 90 when converting from quat -> euler, and in certain orientations increasing the Y rotation ends up rotating the object the opposite direction.