Fix scene c# reload issue
This is not a great fix, but works for now. Basically the issue is that if a scene asset is loaded and the c# scripts reload, trying to clone that scene in the future (i.e. opening the scene while it is loaded) will cause a crash because the script instances do not realize that the c# objects have been destroyed. This could be fixed by using an event listener or something