Runtime shader reloading
This also came with a change to how we unload assets. Previously, it was just based on how recently the asset was accessed. However, this led to misleading information (i.e. the case of shaders or textures) where the data (a shared pointer) was being used somewhere else and the asset claimed it was 'unloading' when in reality the data was never being freed. Now, assets can validate whether or not they should actually be unloaded (i.e. refcount is zero) which 1. improves accuracy of the unload system and 2. enables shader reloading because in the past, the manager claimed the shaders were unloaded when they were actually in use but hadn't been accessed since creation time.