C# managed callback refactor
Refactor the way we import managed callbacks into c#. Originally, we used DllImport to pull functions from native code. However, this was not working on android, and will not be scalable for the future. Now, we use codegen magic and pass an array of function pointers to c# at startup instead.